refactor: ProductRepository & ProductPriceListRepository & ProductImageRepository
