+++ src/Persistence/Persistence/Aggregates/Products/ProductRepository.cs
@@ -7,39 +7,9 @@ namespace Persistence.Aggregates.Products;
 
 public partial class ProductRepository(UniBazzarContext uniBazzarContext) : IProductRepository
 {
-    public void AddProductImage(ProductImage productImage)
-    {
-        uniBazzarContext.Add(productImage);
-    }
-
-    public Task<List<ProductImage>> GetAllProductImagesAsync()
-    {
-        return uniBazzarContext.ProductImages.ToListAsync();
-    }
-
-    public Task<List<ProductImage>> GetImageByProductIdAsync(Guid id)
-    {
-        return uniBazzarContext.ProductImages.Where(x => x.ProductId == id).ToListAsync();
-    }
-
-    public async Task<ProductImage> GetProductImageAsync(Guid id)
-    {
-        var productimage = await uniBazzarContext.ProductImages.FirstOrDefaultAsync(x => x.Id == id);
-        return productimage ?? new ProductImage();
-    }
-
-    public void Remove(ProductImage productImage)
-    {
-        uniBazzarContext.ProductImages.Remove(productImage);
-    }
-
 	public void AddProduct(Product entity)
 	{
 		uniBazzarContext.Add(entity);
-	}
-	public void AddProductPriceList(ProductPriceList productPriceList)
-	{
-        uniBazzarContext.Add(productPriceList);
 	}
 
 	public Task<List<Product>> GetAllProductsAsync()
@@ -53,13 +23,6 @@ public partial class ProductRepository(UniBazzarContext uniBazzarContext) : IPro
 						   .ToListAsync();
 	}
 
-    public async Task<ProductPriceList> GetProductPriceListAsync(Guid id)
-	{
-
-        var productpricelist = await uniBazzarContext.ProductPriceLists.FirstOrDefaultAsync(x => x.Id == id);
-        return productpricelist ?? new ProductPriceList();
-	}
-
 	public async Task<Product> GetProductAsync(Guid id)
     {
 		var product = await uniBazzarContext.Products
@@ -67,23 +30,6 @@ public partial class ProductRepository(UniBazzarContext uniBazzarContext) : IPro
 		return product ?? new Product();
 
 	}
-
-	public Task<List<ProductPriceList>> GetAllProductPriceListAsync()
-	{
-
-        return uniBazzarContext.ProductPriceLists.ToListAsync();
-    }
-
-    public async Task<ProductPriceList> GetPriceByProductId(Guid id)
-    {
-        var productpricelist = await uniBazzarContext.ProductPriceLists.FirstOrDefaultAsync(x => x.ProductId == id);
-        return productpricelist ?? new ProductPriceList();
-	}
-
-	public void Remove(ProductPriceList productPriceList)
-	{
-        uniBazzarContext.ProductPriceLists.Remove(productPriceList);
-    }
 
 	public void RemoveProduct(Product entity)
 	{
